requirements.txt
cycler==0.10.0
helper==2.4.2
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
six==1.11.0
torch==0.4.1
torchvision==0.2.1
